Add query intent features
This is a work in progress branch that adds query intent
features to the ranking model.

These features did not improve our offline ranking metrics
such as NDCG. This suggests that the features did not help
the model, as it began to overfit to the training data sooner.

It's possible these features (or features like these) will help.
We can return to this branch in the future.

https://trello.com/c/BlPOaYXB/1367-add-more-query-intent-features