Paginate over BigQuery query_job response
BigQuery limits the number of rows returned in a query response
to 100,000. This is a problem if we want to generate a training
dataset on over a month of data, since we generate substantially
more data than this.

This commit resolves the issue by using a query job and then
paginating over the 100,000 row data responses. This will ensure
we get all of the data.

Still to do: see what the effect of paginating will be on search
machine memory - we may want to return a lazy enumerator for
queries.

https://trello.com/c/86egVEqC/1392