(wip) Use jsonnet to generate concourse pipeline
The Concourse yaml file is very verbose and has a lot of repetition,
which means:

- if something needs to change (eg, we need a new environment variable
  in all the fetch jobs) then we have to make the same change in several
  places.

- if we want to add deploy jobs for a new model variant, we need to
  copy many lines of yaml when only a couple of lines per environment
  are relevant.

In addition, changes to the pipeline config (this file) and the task
scripts (the .py files) are not really connected at all: the tasks
reference whatever scripts are on 'master' when the task runs.

This commit changes the pipeline to be written in jsonnet.  We already
use jsonnet for generating the content schemas so there is already
knowledge of this in GOV.UK.

Changing all jobs of the same type now requires just changing the
function which generates the job config.  Adding a new deploy job is a
one-line addition.

(todo) And the Python scripts are embedded in the pipeline config,
rather than read from github when the job runs.